[LowerToHW] Revert a1b7ca605 '%0d' default
Revert a change introduced in a1b7ca605 that would cause all FIRRTL radix
substitutions to get an automatic `%0` width specifier.  While this is a
sane default, this exposed situations where users were relying on this
behavior as an actual interface.  I.e., while this was brought up as a
stylistic concern, it was also a functional breakge.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>