[LLHD] Update LowerProcesses pass to work with process results
Add an update version of the process lowering pass. This new version
expects all probes and drives to have been hoisted out of processes. The
pass then converts an `llhd.process` op to an `scf.execute_region` if
the process

- has a single wait op;
- has a simple control flow loop through that single wait op;
- has no destination operands on the wait op; and
- has all operands that are used in the process but defined outside of
  the process observed by the wait op.

This essentially lowers any combinational LLHD processes with a CFG loop
through a wait op which triggers when any of the inputs change into a
simple SCF execute region without the CFG loop. The assumption here is
that the execute region will be re-evaluated if any of the values used
within it change.

Life would be a lot easier if we didn't have to chase through the
observed values of the wait op for this lowering. Ideally LLHD would
have a simplified version of a process that we can canonicalize to early
on.