[FIRRTL] Add VersionAttr
Add a `VersionAttr` which can be used to store the version of a specific
FIRRTL circuit.  While this was previously only necessary to be known
during parsing, due to planned changed in annotation handling, this
information will need to be put on the circuit and preserved through to
`LowerAnnotation`.  This is the first infrastructural piece necessary to
work towards this.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>