[FIRRTL] Add version to CircuitOp
Add a new attribute to FIRRTL's `CircuitOp` that contains the version that
the FIRRTL circuit was parsed as.  This is intended to be static
information that will change compilation behavior.  At present, this is
only intended to be used to change how annotations are parsed.  However,
this is general infrastructure that could be used to make actual
breaking compilation changes if we ever needed it.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>