Update memref.subview handling.
Finally, there was a change to the memref.subview verifier to actually
check bounds when statically possible. Since memref.subview is
technically supposed to "represent a reduced-size view of the
original memref", I think our previous lowering was technically
supposed to be interpreted as out of bound accesses.

In any case, it seems this is the expected use for
memref.collapse_shape, so this was adopted instead of memref.subview
for this use case.