Bump LLVM to 9344b2196cbc36cdc577314bbb2b889606ba6820. (#8404)
There were series of patches to llvm::EquivalenceClasses this week,
and we updated accordingly:

* There is no longer a template parameter for the comparator
* The idiom for iterating over EquivalenceClasses was updated
* The idiom for iterating over members was updated

There was a change to the InlinerUtils functions to now include a
required callback for ops that are cloned. To update our call sites,
this default constructs an InlinerConfig and uses the default
clone callback.

Finally, there was a change to the memref.subview verifier to actually
check bounds when statically possible. Since memref.subview is
technically supposed to "represent a reduced-size view of the
original memref", I think our previous lowering was technically
supposed to be interpreted as out of bound accesses.

In any case, it seems this is the expected use for
memref.collapse_shape, so this was adopted instead of memref.subview
for this use case.