[circt-bmc] Correct signature of generated `printf`
`lib/Tools/circt-bmc/LowerToBMC.cpp` generates a `printf`
stub with (previously) the signature `(ptr, ...) -> void`.
`lib/Conversion/SMTToZ3LLVM/LowerSMTToZ3LLVM.cpp` looks up a function
named `printf`, and expects the signature to be `(ptr, ...) -> i32`.
Because these types are not the same, `circt-bmc` crashes (but only when
invoked with `--print-solver-output`).

Because the `(ptr, ...) -> i32` signature is correct, this changes the
generated stub to have that type.