[FIRRTL] Dedup: dedup wires when only one has a symbol
This change allows deduplication of wires when only one of them has an
inner symbol. Before, when hashing an operation we would include the
index/id of the inner symbol in the wire's hash, which was done to make
sure that local inner symbol references were targeting the same
declaration between two modules. Now, we preprocess the inner symbols
inside a module, and with a stable identity of inner symbols we can
confidently use the index of an innersymbol when hashing an inner symbol
reference. With this change, we don't have to include the inner symbol
on the declaration operation, meaning that if there are no references to
the symbol, the fact that an declaration has an inner symbol does not
matter at all.