[hw] Add second insertPoint API to HierPathBuilder
Add a second API to `hw::HierPathBuilder` which can be used if the user is
maintaining the insertion point themselves.  This is generally better for
stability/determinism as the insertion point can then be made just above a
FIRRTL module when doing work in a parallel region operating on FIRRTL
modules independently.

Rewrite the original, narrower API which uses the global insertion point
to use the new, wider API.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>