[verilator driver] Fix for clang
Was failing on link due to Verilator trace library not being found. I
think this is because Verilator does not produce this library unless
tracing is enabled. Further, clang (by default) doesn't detect that the
symbols aren't actually needed (ie it doesn't optimize away the 'if'
branch) so it requires them.