[OM] Ensure every sublist is finished in ListConcatOp evaluation.
We would previously check if the value corresponding to each operand
of the ListConcatOp was finished evaluating, but in the case where the
operand is a ReferenceValue, the reference may be done evaluating
before while the underlying ListValue is not.

To handle this, a check was added that the result of extractList is
fully evaluated, and the evaluation of the ListConcatOp doesn't
proceed until that is true.

A small unit test shows this scenario, where there are object
references in lists passed as field references to a ListConcatOp.