[FIRRTL] Update blackbox metadata to include "in DUT" logic.
We are porting this metadata from a JSON output to use the FIRRTL
object operations. As part of this, we need to distinguish if each
object represents a blackbox that is used in the DUT or not, using the
same logic to split the blackboxes into two different output files in
the old JSON version.