[PyCDE] Add `Bundle.create_uturn` method
reates two bundle signals which talk to each other. The types of them
are the inverse of each other, the first one matching this type.  E.g.
anything which is sent on the TO channel 'foo' on the first channel will
be transmitted to the FROM channel 'foo' on the second bundle.