[LLHD] Fix use-after-free in Mem2Reg, add missing reaching def for probe
Fix an issue in Mem2Reg where the erasing of ops could potentially lead
to a use-after-free issue when that op was also in the
`opsToEraseIfUnused` set. A new `eraseOp` helper now ensures that the op
is removed from the set before erasing it. This issue doesn't show up
today since Mem2Reg does not support drives to projections of signals,
which would create sufficiently complex new operations where this issue
could show up.

Also make probes insert a reaching definition for the delayed version of
a slot. This is _technically_ only correct if the aggregates are packed.
It's safe to assume they all are for now, since unpacked aggregates are
separated by the SROA pass. However, we'll want to fix this in the
future by providing a way for drives to have a mask indicating which
subfields of a signal to update.