[LLHD] Add better LLHD dialect doc, refactor op tablegen includes
Add a dedicated `LLHD.md` dialect documentation that acts as a rationale
and which pulls in generated documentation for the types, attributes,
ops, and passes of the dialect.

This requires bundling the LLHD ops in an `LLHDOps.td` file which
contains includes for the files that contain the actual op definitions.