[FIRRTL] Add moveDut field for inject DUT hier
Add a new field to the `InjectDUTHierarchyAnnotation` called `moveDut`.
This is used to control the placement of the `MarkDUTAnnotation` after the
pass runs.  This is done becasue we are exploring an internal flow which
wants to change the behavior of this pass to work this way.  We would like
to make this the default behavior, but will initially guard it with this
annotaiton option.

This patch is backwards compatible.  It will default to the old behavior
if it does not see the new field in the annotation.

This makes no changes to the placement of _other_ annotations.  I.e., if
there are annotations that were originally on the DUT, then those will
still be moved.  I went back and forth on this.  However, it shouldn't
really matter where _other_ annotations are nor should it matter where
NLAs are rooted so long as they are information preserving.  This
additionally makes the pass easier to maintain as it doesn't require
having different NLA movement logic for different annotation settings.

CC: @azidar

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>