[LLHD] Add llhd.sig.extract support to Mem2Reg
Extend LLHD's Mem2Reg pass to support `llhd.sig.extract` projections. To
implement the extraction and injection of bit ranges at dynamic offsets,
add a few helper functions to `CombOps.h` that assemble the necessary
operations. We may want to add a `comb.inject` op in the future to
simplify this.