[LLHD] Hoist probes out of scf.if ops
Extend the HoistSignals pass to also hoist probes out of `scf.if` ops
where possible. The Moore dialect generates a bunch of `scf.if` ops for
ternary operations like `x ? doA() : doB()` when it is not trivially
clear that `doA()` and `doB()` have no side-effects. We'll need a better
way to deal with this at some point.