[hw] Add hw.donttouch operation
Add a `hw.donttouch` operation which is an inner symbol user that
indicates that the inner symbol should not be optimized, removed, or
otherwise touched by the compiler.

This is added to fill a hole where unused inner symbols are being used to
mean "don't touch".  The lack of a user means that all inner symbols are
effectively "don't touch" when only certain ones should be.  Concretely,
this is being added as a step towards better optimization of read probes.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>