[FIRRTL] Dedup: MustDedup support merging of pub+priv modules (#8546)
We upgraded dedup to be able to merge modules where one is public, as
long as the other modules were private.  The MustDedup error message was
not updated to handle this change in behaviour, and still assumed that
it was impossible to merge any public module with other modules.  This
changes MustDedup to only detect this as a difference if neither module
is marked as private.