[LLHD] Add combinational process op (#8536)
Add a dedicated op for processes that should trigger when any of the
values used as operands in its body region change. This reflects the
behavior of `always_comb` in SystemVerilog and will allow passes such as
desequentialization and process lowering to convert `llhd.process`es to
this more streamlined and specialized representation.

A later commit will replace uses of `scf.execute_region` in the LLHD
pass pipeline with `llhd.combinational`.