[FIRRTL] Extract instances above DUT w/ moveDut
Modify newly added behavior related to the `moveDut` parameter added to
`InjectDUTHierarchyAnnotation` [[1]].  Change the behavior so, when the
`moveDut` parameter in the newly preserved annotation, instances will be
extracted one level above the newly _moved_ DUT.

This adds tech debt as it repurposes instance extraction to allow for both
hierarchy injection and extraction beyond the DUT.  Previously, these were
intended to be mutually exclusive.

The tech debt is acceptable because @mwachs5 is planning to remove this
flow entirely.

[1]: 91d94336afb65330a3b9f265aa2ad3d463a19a07

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@sifive.com>