[ESI] Set mode before `connectImpl`
`connectImpl` is shared across all connect functions. However, if `mode` is only set _after_ `connectImpl` returns, `connectImpl` doesn't have a direct way to check the connect mode (we could look for whether or not the callback has been assigned, but that's then also placing an assumption on the ordering of things in `Ports.cpp`).