[circt-verilog] Add updated LLHD passes to the pipeline (#8579)
Enable the LLHD passes that have been updated in the past few months in
circt-verilog's pass pipeline. This contains revised versions of the
Mem2Reg, Deseq, and LowerProcesses passes, alongside a new HoistSignals
pass and the CombineDrives pass factored out of Sig2Reg.

This also disables the SROA pass on LLHD, moving it to only operate on
Moore modules. @maerhart has been looking into fixing an upstream bug in
SROA when it's used on operations in graph regions. Until that is fixed,
we can just skip SROA on HW modules for now.

Also add an end-to-end test for circt-verilog to ensure that the basic
forms of registers get detected and lowered properly.