[Kanagawa] Replace CSE with specialized pass (#8599)
Running CSE on a whole design is neither necessary nor advisable. It forces all the dialects used to support CSE but at least one (which we use) has issues with it. So we add a PR which targets redundant operations in the `kanagawa` dialect which passes assume to be non-redundant.

Also updates `kanagawatool` to add new passes which are now necessary.