[FIRRTL] FIRParser: support caching constants in match statements
This fixes a long standing bug when parsing match statements which
contain numeric constants.

The FIRParser has a contant cache so that we can reuse any constant
operation created as we parse, by inserting them into the top-level body
of the module.  To find the top level block, we were searching upwards
from the current insertion point.

In general, when creating any operation, we are required to know how
many regions the operation has; it is impossible to add more regions
after creating the operation. Due to this, when parsing the arms of a
match statement, the expressions are inserted into a region which is not
attached to any operation.  This means we can not search upwards from
the insertion point to find the top level block. When operations do not
have this problem as they always have space 2 regions.

This change passes in the top level block when creating the
FIRModuleContext.  If it turns out that the region we are inserting into
is not underneath the module, then we can insert the constant at the end
of the module's block.

This change also makes the cache itself a private member, and removes a
bit of code which was unnecessarily directly accessing the cache.