[LLHD] Use initial signal value to fill gaps in CombineDrives
If the CombineDrives pass can see and reason about all uses of a signal,
use the signal's initial value to fill in any undriven parts of the
signal. This works around a somewhat common pattern in input Verilog
where an array signal has some of its elements completely unused --
neither read nor written:

```systemverilog
logic [1:0][31:0] a;
assign a[0] = ...;
assign ... = a[0];
// a[1] is completely unused
```

Driving previously undriven parts of the signal with their initial value
is valid if we can guarantee that there are truly no drivers to those
parts. In that case, adding a driver does not cause any drive conflicts,
and the driven value is the same as the value already on the signal,
namely the initial value.

One caveat is that this transform is only valid if the initial value of
the signal is a constant. This is not always guaranteed, with the
initial value potentially being provided by a constant passed in through
a port. The Deseq pass faces a similar issue where real-world Verilog
usage makes use of somewhat unclear semantics in initialization order.
We should revisit the `llhd.sig` op at a later stage and consider using
@uenoku's `seq.initial` mechanism to ensure that the value is static
during runtime, or make the signal initial value an attribute.