[circt-verilog] Replace arith.select with comb.mux
Add a small pass to the LLHD dialect to convert `arith.select` ops to
`comb.mux` operations. (Also `arith.constant` to `hw.constant`.) The
canonicalizers in the CF dialect produce `arith.select` operations, but
we would like the circt-verilog pipeline to produce HW/Comb/Seq output
where possible.

In the future we might want to create a full Arith-to-Comb conversion
pass, similar to the Comb-to-Arith we already have. This feels like a
bit of overkill to do just for the select ops, but we migth run into
more of these ops at a later point.